+++ distribution/src/resources/sql/20171229-13-update_point_ajout_argent.sql
@@ -1,2 +1,2 @@
 ALTER TABLE point
-ADD COLUMN argent bigint;
+ADD COLUMN argent numeric(12,2);
